Fix auto-update and add progress bar on download. (Just need someone to test...)
Auto update works as follow:
Current_Release folder in svn has 2 files:
current.dol and current.txt
current.dol is the new dol and current.txt describes it (its used to see if the version available is newer).

I also added an autoUpdate setting in the xml file we need a gui option for... it lets you set wether or not you get prompted when an update is available.